requirements.txt
aiodns==3.5.0
aiofile==3.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.16.4
anyio==4.11.0
asgiref==3.11.0
attrs==25.4.0
blinker==1.9.0
caio==0.9.24
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.3.1
Django==6.0
django-filter==25.2
djangorestframework==3.16.1
drf-yasg==1.21.14
ecdsa==0.19.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.6
Mako==1.3.10
Markdown==3.10.1
MarkupSafe==3.0.2
multidict==6.7.0
numpy==2.3.5
openapi-codec==1.3.2
packaging==25.0
pandas==2.3.3
pillow==12.1.0
pluggy==1.6.0
propcache==0.4.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycares==4.11.0
pycparser==2.23
Pygments==2.19.2
PyJWT==2.9.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.3
rsa==4.9.1
setuptools==80.9.0
shtab==1.7.2
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.4
stream-chat==4.26.0
stripe==11.3.0
svix==1.32.0
termcolor==3.1.0
types-Deprecated==1.3.1.20251101
types-python-dateutil==2.9.0.20251108
typing_extensions==4.14.1
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==2.0.1
yarl==1.22.0

